cure_seq/requirements.txt
torch>=2.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate
Pillow
torchvision
